Deleted plexi.cs from INFOIBV, it is already being referenced to by form.cs. Moved all remaining processors like Threshold and Grayscale from the form.cs code to plexi\program.cs
