+++ INFOIBV/Form1.cs
@@ -31,48 +31,7 @@ namespace INFOIBV
             }
         }
 
-        public class Negative : Processor
-        {
-            public override Color Transform(Color c)
-            {
-                return Color.FromArgb(255 - c.R, 255 - c.G, 255 - c.B);
-            }
-        }
-
-        public class RotateRight : Processor
-        {
-            public override Color[,] Process(Color[,] image)
-            {
-                int w = image.GetLength(0), h = image.GetLength(1);
-                var newImage = new Color[h, w];
-                for (var x = 0; x < w; x++)
-                    for (var y = 0; y < h; y++)
-                        newImage[y, x] = image[x, h - y - 1];
-                return newImage;
-            }
-        }
-
-        public class Grayscale : Processor
-        {
-            public override Color Transform(Color c)
-            {
-                int grayVal = (int)((c.R * 0.21) + (c.G * 0.72) + (c.B * 0.07)); // The numbers represent a weight for better grayscale values
-                return Color.FromArgb(grayVal, grayVal, grayVal);
-            }
-        }
-
-        public class Threshold : Processor
-        {
-            public override Color Transform(Color c)
-            {
-                int t = 128;    // Threshold value
-
-                if (c.R > t) { return Color.White; }
-                else return Color.Black;
-            }
-        }
-
-        private void applyButton_Click(object sender, EventArgs e)
+       private void applyButton_Click(object sender, EventArgs e)
         {
             if (InputImage == null) return;                                 // Get out if no input image
             pictureBox2.Image = null;                                       // Clear output image
